# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MagicMirror Remote Control API",
    "description": "The MMM-Remote-Control Module for MagicMirror² implements a RESTful(-ish) API to control the Magic Mirror using the existing functionality built-in to MMM-Remote-Control, as well as the notifications commands built into most modules. In addition, the API creates a basic framework which allows for each module to expand or customize their own API by a simple notification.",
    "contact": {
      "name": "MMM-Remote-Control",
      "url": "https://github.com/Jopyth/MMM-Remote-Control"
    },
    "license": {
      "name": "MIT License",
      "url": "https://github.com/Jopyth/MMM-Remote-Control/blob/master/LICENSE"
    },
    "version": "2.3.0+"
  },
...
  "paths": {
...
    "/api/saves": {
      "get": {
        "tags": [
          "API"
        ],
        "summary": "Return backups of the config file",
        "security": [
          {
            "apikey": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "query": {
                    "data": "saves"
                  },
                  "data": [
                    "2020-11-14T20:41:52.589Z",
                    "2020-11-14T20:41:13.876Z",
                    "2020-11-09T20:38:21.783Z",
                    "2020-11-09T20:32:59.994Z"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/classes/{:value}": {
      "get": {
        "tags": [
          "Module Control",
          "API"
        ],
        "summary": "Show or use a class",
        "parameters": [
          {
            "name": ":value",
            "in": "path",
            "required": true,
            "allowEmptyValue": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "apikey": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "query": {
                    "data": "classes"
                  },
                  "data": {
                    "test": {
                      "toggle": "calendar"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Invalid value test provided in request. Use /api/classes to see actual values"
                }
              }
            }
          }
        }
      }
    },
    "/api/command/{:value}": {
      "get": {
        "tags": [
          "Mirror Control",
          "API"
        ],
        "summary": "Execute command saved in config",
        "parameters": [
          {
            "name": ":value",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "apikey": []
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Value is needed"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Forbidden: API Key Not Provided in Config! Use secureEndpoints to bypass this message"
                }
              }
            }
          }
        }
      }
    },
    "/api/config": {
      "get": {
        "tags": [
          "API"
        ],
        "summary": "Return the current MagicMirror configuration",
        "security": [
          {
            "apikey": []
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "query": {
                    "data": "config"
                  },
                  "data": {
                    "address": "0.0.0.0",
                    "electronOptions": {
                      "fullscreen": false
                    },
                    "ipWhitelist": [],
                    "modules": [
                      {
                        "module": "MMM-Remote-Control",
                        "position": "top_right",
                        "config": {
                          "showModuleApiMenu": true
                        }
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/config/edit": {
      "post": {
        "tags": [
          "Mirror Control",
          "API"
        ],
        "summary": "Modify the current MagicMirror configuration",
        "security": [
          {
            "apikey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "query": {
                    "data": "config"
                  },
                  "data": {
                    "address": "0.0.0.0",
                    "electronOptions": {
                      "fullscreen": false
                    },
                    "ipWhitelist": [],
                    "modules": [
                      {
                        "module": "MMM-Remote-Control",
                        "position": "top_right",
                        "config": {
                          "showModuleApiMenu": true
                        }
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Invalid URL provided in request body"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Forbidden: API Key Not Provided in Config! Use secureEndpoints to bypass this message"
                }
              }
            }
          }
        }
      }
    },
    "/api/monitor/{:action}": {
      "get": {
        "tags": [
          "Mirror Control",
          "API"
        ],
        "summary": "Control the connected display, turning it on or off.",
        "parameters": [
          {
            "name": ":action",
            "in": "path",
            "required": true,
            "allowEmptyValue": true,
            "schema": {
              "type": "string"
            },
            "examples": {
              "on": {
                "summary": "Turn on",
                "value": "on"
              },
              "off": {
                "summary": "Turn off",
                "value": "off"
              },
              "toggle": {
                "summary": "Toggle",
                "value": "toggle"
              },
              "status": {
                "summary": "Status",
                "value": "status"
              }
            }
          }
        ],
        "security": [
          {
            "apikey": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "monitor": "on"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/minimize": {
      "get": {
        "tags": [
          "Mirror Control",
          "API"
        ],
        "summary": "Minimize the browser instance (Electron)",
        "security": [
          {
            "apikey": []
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "success": true
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Forbidden: API Key Not Provided in Config! Use secureEndpoints to bypass this message"
                }
              }
            }
          }
        }
      }
    },
    "/api/togglefullscreen": {
      "get": {
        "tags": [
          "Mirror Control",
          "API"
        ],
        "summary": "Toggle fullscreen (Electron)",
        "security": [
          {
            "apikey": []
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "success": true
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Forbidden: API Key Not Provided in Config! Use secureEndpoints to bypass this message"
                }
              }
            }
          }
        }
      }
    },
    "/api/devtools": {
      "get": {
        "tags": [
          "Mirror Control",
          "API"
        ],
        "summary": "Toggle devtools (Electron)",
        "security": [
          {
            "apikey": []
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "success": true
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Forbidden: API Key Not Provided in Config! Use secureEndpoints to bypass this message"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/save": {
      "get": {
        "tags": [
          "API"
        ],
        "summary": "Saves the MMM-Remote-Control defaults to a file.",
        "security": [
          {
            "apikey": []
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "example": null
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Forbidden: API Key Not Provided in Config! Use secureEndpoints to bypass this message"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apikey": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}